+++ src/main/java/br/com/zitrus/procedurecontrol/dao/ProcedureDao.java
@@ -1,2 +1,12 @@
-package br.com.zitrus.procedurecontrol.dao;public interface ProcedureDao {
+package br.com.zitrus.procedurecontrol.dao;
+
+import br.com.zitrus.procedurecontrol.model.Proceduresql;
+
+import java.util.List;
+
+public interface ProcedureDao {
+
+    List<Proceduresql> getAllProcedure();
+
+    Proceduresql findProcedureById(Long id);
 }