+++ src/main/java/br/com/zitrus/procedurecontrol/dao/RuleDao.java
@@ -1,0 +1,8 @@
+package br.com.zitrus.procedurecontrol.dao;
+
+import br.com.zitrus.procedurecontrol.model.Authorization;
+
+public interface RuleDao {
+
+    boolean validateAuthorization(Authorization authorization);
+}