+++ src/main/java/br/com/zitrus/procedurecontrol/model/Authorization.java
@@ -2,7 +2,13 @@ package br.com.zitrus.procedurecontrol.model;
 
 import br.com.zitrus.procedurecontrol.enums.AuthorizationEnum;
 import br.com.zitrus.procedurecontrol.enums.GenderEnum;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@Getter
+@Setter
+@NoArgsConstructor
 public class Authorization {
 
     private Long id;
@@ -26,56 +32,6 @@ public class Authorization {
         this.proceduresql = proceduresql;
         this.authorization = authorization;
         this.age = age;
-        this.gender = gender;
-    }
-
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Proceduresql getProcedure() {
-        return proceduresql;
-    }
-
-    public void setProcedure(Proceduresql proceduresql) {
-        this.proceduresql = proceduresql;
-    }
-
-    public AuthorizationEnum getAuthorization() {
-        return authorization;
-    }
-
-    public void setAuthorization(AuthorizationEnum authorization) {
-        this.authorization = authorization;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public GenderEnum getGender() {
-        return gender;
-    }
-
-    public void setGender(GenderEnum gender) {
         this.gender = gender;
     }
 }